// src/@types/Entities/BetOverlay.d.ts
export interface BetOverlay {
    userId: number;
    fontFamily: string;
    fontVariant: string;
    distributionBackground: string;
    distributionFont: string;
    distributionFontSize: string;
    distributionColorLeft: string;
    distributionColorRight: string;
    timerBackground: string;
    timerFont: string;
    timerFontSize: string;
    toplistBackground: string;
    toplistFont: string;
    toplistFontSize: string;
    toplistShowRank: boolean;
    toplistShowTotalBets: boolean;
    toplistShowAccuracy: boolean;
}